#8 Small restructuring and updated description of pubspec.yaml
